requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic>=2
sqlalchemy[asyncio]
asyncpg
psycopg[binary]
python-dotenv
loguru
pyyaml
